bot/requirements.txt
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.15.0
idna==2.8
oauthlib==3.1.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
python-telegram-bot==12.1.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
SQLAlchemy==1.3.12
tornado==6.0.3
uritemplate==3.0.1
urllib3==1.25.7
python-dotenv==0.11.0


